// frontend/src/types/store.ts
import type { Friend, User } from '@/types/user';
import type { Conversation, Message } from '@/types/chat';
import { Socket } from 'socket.io-client';
import type { FriendRequest } from '@/types/user';

export interface AuthState {
    accessToken: string | null;
    user: User | null;
    loading: boolean;

    setAccessToken: (accessToken: string) => void;

    setUser: (user: User) => void;

    clearState: () => void;

    signUp: (
        username: string,
        password: string,
        email: string,
        firstName: string,
        lastName: string,
    ) => Promise<void>;

    signIn: (username: string, password: string) => Promise<void>;

    signOut: () => Promise<void>;

    fetchMe: () => Promise<void>;

    refresh: () => Promise<void>;
}

export interface ThemeState {
    isDark: boolean;
    toggleTheme: () => void;
    setTheme: (dark: boolean) => void;
}

export interface ChatState {
    conversations: Conversation[];
    messages: Record<
        string,
        {
            items: Message[];
            hasMore: boolean; // infinite-scroll
            nextCursor?: string | null; //pagination
        }
    >;
    activeConversationId: string | null;
    convoLoading: boolean;
    messageLoading: boolean;
    loading: boolean;
    reset: () => void;

    setActiveConversation: (id: string | null) => void;
    fetchConversations: () => Promise<void>;
    fetchMessages: (conversationId?: string) => Promise<void>;
    sendDirectMessage: (
        recipientId: string,
        content: string,
        imgUrl?: string,
    ) => Promise<void>;
    sendGroupMessage: (
        conversationId: string,
        content: string,
        imgUrl?: string,
    ) => Promise<void>;
    // add message
    addMessage: (message: Message) => Promise<void>;
    //update conversation
    updateConversation: (conversation: unknown) => void;
    // mark conversation as seen
    markAsSeen: () => void;
    // add convo:
    addConvo: (convo: Conversation) => void;
    createConversation: (
        type: 'direct' | 'group',
        name: string,
        memberIds: string[],
    ) => Promise<void>;
}

export interface SocketState {
    socket: Socket | null;
    onlineUsers: string[];
    connectSocket: () => void;
    disconnectSocket: () => void;
}

export interface FriendState {
    friends: Friend[];
    loading: boolean;
    receivedList: FriendRequest[];
    sentList: FriendRequest[];
    searchByUsername: (username: string) => Promise<User | null>;
    addFriend: (to: string, message?: string) => Promise<string>;
    getAllFriendRequests: () => Promise<void>;
    acceptFriendRequest: (requestId: string) => Promise<void>;
    declineFriendRequest: (requestId: string) => Promise<void>;
    getFriends: () => Promise<void>;
}

export interface UserState {
    updateAvatarUrl: (formData: FormData) => Promise<void>;
}